implementado guardar partida

+++ spring-boot-proyecto/src/main/java/com/proyecto/bombsaway/entidades/EntidadArtilleria.java
@@ -16,13 +16,13 @@ public class EntidadArtilleria {
 	private EntidadJugador jugador;
 
 	@Column(name = "eje_x", nullable = false)
-	private int ejeX;
+	private double ejeX;
 
 	@Column(name = "eje_y", nullable = false)
-	private int ejeY;
+	private double ejeY;
 
 	@Column(name = "angulo", nullable = false)
-	private int angulo;
+	private double angulo;
 
 	@Column(name = "destruida", nullable = false)
 	private boolean destruida;
@@ -33,7 +33,8 @@ public class EntidadArtilleria {
 	public EntidadArtilleria() {
 	}
 
-	public EntidadArtilleria(int id, EntidadJugador jugador, int ejeX, int ejeY, int angulo, boolean destruida, boolean visible) {
+	public EntidadArtilleria(int id, EntidadJugador jugador, double ejeX, double ejeY, double angulo, boolean destruida,
+			boolean visible) {
 		this.id = id;
 		this.jugador = jugador;
 		this.ejeX = ejeX;
@@ -59,27 +60,27 @@ public class EntidadArtilleria {
 		this.jugador = jugador;
 	}
 
-	public int getEjeX() {
+	public double getEjeX() {
 		return ejeX;
 	}
 
-	public void setEjeX(int ejeX) {
+	public void setEjeX(double ejeX) {
 		this.ejeX = ejeX;
 	}
 
-	public int getEjeY() {
+	public double getEjeY() {
 		return ejeY;
 	}
 
-	public void setEjeY(int ejeY) {
+	public void setEjeY(double ejeY) {
 		this.ejeY = ejeY;
 	}
 
-	public int getAngulo() {
+	public double getAngulo() {
 		return angulo;
 	}
 
-	public void setAngulo(int angulo) {
+	public void setAngulo(double angulo) {
 		this.angulo = angulo;
 	}
