+++ spring-boot-proyecto/src/main/java/com/proyecto/bombsaway/entidades/EntidadAvion.java
@@ -28,13 +28,13 @@ public class EntidadAvion {
 	private EntidadJugador jugador;
 
 	@Column(name = "eje_x", nullable = false)
-	private int ejeX;
+	private double ejeX;
 
 	@Column(name = "eje_y", nullable = false)
-	private int ejeY;
+	private double ejeY;
 
 	@Column(name = "angulo", nullable = false)
-	private int angulo;
+	private double angulo;
 
 	@Column(name = "estado", nullable = false)
 	@Enumerated(EnumType.STRING)
@@ -55,8 +55,8 @@ public class EntidadAvion {
 	public EntidadAvion() {
 	}
 
-	public EntidadAvion(int id, EntidadJugador jugador, int ejeX, int ejeY, int angulo, EstadoAvion estado, int vida, int combustible,
-			boolean tieneBomba, boolean visible) {
+	public EntidadAvion(int id, EntidadJugador jugador, double ejeX, double ejeY, double angulo, EstadoAvion estado,
+			int vida, int combustible, boolean tieneBomba, boolean visible) {
 		this.id = id;
 		this.jugador = jugador;
 		this.ejeX = ejeX;
@@ -85,27 +85,27 @@ public class EntidadAvion {
 		this.jugador = jugador;
 	}
 
-	public int getEjeX() {
+	public double getEjeX() {
 		return ejeX;
 	}
 
-	public void setEjeX(int ejeX) {
+	public void setEjeX(double ejeX) {
 		this.ejeX = ejeX;
 	}
 
-	public int getEjeY() {
+	public double getEjeY() {
 		return ejeY;
 	}
 
-	public void setEjeY(int ejeY) {
+	public void setEjeY(double ejeY) {
 		this.ejeY = ejeY;
 	}
 
-	public int getAngulo() {
+	public double getAngulo() {
 		return angulo;
 	}
 
-	public void setAngulo(int angulo) {
+	public void setAngulo(double angulo) {
 		this.angulo = angulo;
 	}
