+++ spring-boot-proyecto/src/main/java/com/proyecto/bombsaway/entidades/EntidadBala.java
@@ -12,13 +12,13 @@ public class EntidadBala {
 	private int id;
 
 	@Column(name = "eje_x", nullable = false)
-	private int ejeX;
+	private double ejeX;
 
 	@Column(name = "eje_y", nullable = false)
-	private int ejeY;
+	private double ejeY;
 
 	@Column(name = "angulo", nullable = false)
-	private int angulo;
+	private double angulo;
 
 	@Column(name = "visible", nullable = false)
 	private boolean visible;
@@ -26,7 +26,7 @@ public class EntidadBala {
 	public EntidadBala() {
 	}
 
-	public EntidadBala(int id, int ejeX, int ejeY, int angulo, boolean visible) {
+	public EntidadBala(int id, double ejeX, double ejeY, double angulo, boolean visible) {
 		this.id = id;
 		this.ejeX = ejeX;
 		this.ejeY = ejeY;
@@ -42,27 +42,27 @@ public class EntidadBala {
 		this.id = id;
 	}
 
-	public int getEjeX() {
+	public double getEjeX() {
 		return ejeX;
 	}
 
-	public void setEjeX(int ejeX) {
+	public void setEjeX(double ejeX) {
 		this.ejeX = ejeX;
 	}
 
-	public int getEjeY() {
+	public double getEjeY() {
 		return ejeY;
 	}
 
-	public void setEjeY(int ejeY) {
+	public void setEjeY(double ejeY) {
 		this.ejeY = ejeY;
 	}
 
-	public int getAngulo() {
+	public double getAngulo() {
 		return angulo;
 	}
 
-	public void setAngulo(int angulo) {
+	public void setAngulo(double angulo) {
 		this.angulo = angulo;
 	}
